refactor(server): merge StoreOrderService into OrderService
