+++ server/app/src/main/kotlin/pos/ambrosia/Api.kt
@@ -17,6 +17,7 @@ import io.ktor.server.plugins.origin
 import io.ktor.server.websocket.WebSockets
 import io.ktor.server.websocket.pingPeriod
 import io.ktor.server.websocket.timeout
+import kotlinx.serialization.json.Json
 import org.slf4j.LoggerFactory
 import pos.ambrosia.api.configureAuth
 import pos.ambrosia.api.configureCategories
@@ -58,7 +59,7 @@ class Api {
         AppConfig.loadConfig() // Load the configuration
         val config = environment.config // Configure the application
         handler() // Install exception handlers
-        install(ContentNegotiation) { json() }
+        install(ContentNegotiation) { json(Json { ignoreUnknownKeys = true }) }
         install(CORS) {
             allowCredentials = true
             anyHost()
