+++ server/app/src/main/kotlin/pos/ambrosia/services/OrderService.kt
@@ -5,6 +5,10 @@ import pos.ambrosia.models.Order
 import pos.ambrosia.models.OrderDish
 import pos.ambrosia.models.OrderWithPayment
 import pos.ambrosia.models.OrderWithPaymentFilters
+import pos.ambrosia.models.StoreCheckoutRequest
+import pos.ambrosia.models.StoreCheckoutResponse
+import pos.ambrosia.models.StoreOrder
+import pos.ambrosia.models.StoreOrderItem
 import java.sql.Connection
 import java.sql.PreparedStatement
 import java.util.UUID
@@ -14,38 +18,64 @@ class OrderService(
 ) {
     companion object {
         private const val ADD_ORDER =
-            "INSERT INTO orders (id, user_id, table_id, waiter, status, total, created_at) VALUES (?, ?, ?, ?, ?, ?, ?)"
+            "INSERT INTO orders (id, user_id, table_id, status, total, created_at) VALUES (?, ?, ?, ?, ?, ?)"
         private const val GET_ORDERS =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE is_deleted = 0"
         private const val GET_ORDER_BY_ID =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE id = ? AND is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE id = ? AND is_deleted = 0"
         private const val UPDATE_ORDER =
-            "UPDATE orders SET user_id = ?, table_id = ?, waiter = ?, status = ?, total = ? WHERE id = ?"
+            "UPDATE orders SET user_id = ?, table_id = ?, status = ?, total = ? WHERE id = ?"
         private const val DELETE_ORDER = "UPDATE orders SET is_deleted = 1 WHERE id = ?"
         private const val CHECK_USER_EXISTS = "SELECT id FROM users WHERE id = ? AND is_deleted = 0"
         private const val CHECK_TABLE_EXISTS = "SELECT id FROM tables WHERE id = ? AND is_deleted = 0"
         private const val GET_ORDERS_BY_TABLE =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE table_id = ? AND is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE table_id = ? AND is_deleted = 0"
         private const val GET_ORDERS_BY_USER =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE user_id = ? AND is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE user_id = ? AND is_deleted = 0"
         private const val GET_ORDERS_BY_STATUS =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE status = ? AND is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE status = ? AND is_deleted = 0"
         private const val GET_ORDERS_BY_DATE_RANGE =
-            "SELECT id, user_id, table_id, waiter, status, total, created_at FROM orders WHERE created_at BETWEEN ? AND ? AND is_deleted = 0"
+            "SELECT id, user_id, table_id, status, total, created_at FROM orders WHERE created_at BETWEEN ? AND ? AND is_deleted = 0"
         private const val GET_TOTAL_SALES_BY_DATE =
             "SELECT SUM(total) AS total_sales FROM orders WHERE DATE(created_at) = ? AND status = 'paid' AND is_deleted = 0"
+
+        // ── Store-specific queries ────────────────────────────────────────────
+        private const val STORE_GET_ORDER =
+            "SELECT o.id, o.user_id, u.name as user_name, o.status, CAST(o.total AS INTEGER) as total, o.created_at FROM orders o LEFT JOIN users u ON o.user_id = u.id WHERE o.id = ? AND o.is_deleted = 0 AND o.table_id IS NULL"
+        private const val STORE_GET_ORDERS =
+            "SELECT o.id, o.user_id, u.name as user_name, o.status, CAST(o.total AS INTEGER) as total, o.created_at FROM orders o LEFT JOIN users u ON o.user_id = u.id WHERE o.is_deleted = 0 AND o.table_id IS NULL ORDER BY o.created_at DESC"
+        private const val STORE_GET_ORDERS_BY_STATUS =
+            "SELECT o.id, o.user_id, u.name as user_name, o.status, CAST(o.total AS INTEGER) as total, o.created_at FROM orders o LEFT JOIN users u ON o.user_id = u.id WHERE o.is_deleted = 0 AND o.table_id IS NULL AND o.status = ? ORDER BY o.created_at DESC"
+        private const val STORE_INSERT_CHECKOUT_ORDER =
+            "INSERT INTO orders (id, user_id, table_id, status, total, created_at) VALUES (?, ?, NULL, 'paid', ?, datetime('now'))"
+        private const val STORE_INSERT_ORDER_ITEM =
+            "INSERT INTO order_products (order_id, product_id, quantity, price_at_order) VALUES (?, ?, ?, ?)"
+        private const val STORE_DECREMENT_STOCK =
+            "UPDATE products SET quantity = quantity - ? WHERE id = ? AND is_deleted = 0 AND quantity >= ?"
+        private const val STORE_INSERT_TICKET =
+            "INSERT INTO tickets (id, order_id, user_id, ticket_date, status, total_amount, notes) VALUES (?, ?, ?, datetime('now'), 1, ?, ?)"
+        private const val STORE_INSERT_PAYMENT =
+            "INSERT INTO payments (id, method_id, currency_id, transaction_id, amount) VALUES (?, ?, ?, ?, ?)"
+        private const val STORE_INSERT_TICKET_PAYMENT =
+            "INSERT INTO ticket_payments (payment_id, ticket_id) VALUES (?, ?)"
+        private const val STORE_GET_ITEMS =
+            "SELECT product_id, quantity, price_at_order FROM order_products WHERE order_id = ?"
+        private const val STORE_CANCEL_ORDER =
+            "UPDATE orders SET status = 'closed' WHERE id = ? AND status = 'open' AND table_id IS NULL"
+
         private const val GET_ORDERS_WITH_PAYMENTS_BASE =
             """
             SELECT o.id,
                    o.user_id,
+                   u.name AS user_name,
                    o.table_id,
-                   o.waiter,
                    o.status,
                    o.total,
                    o.created_at,
                    GROUP_CONCAT(DISTINCT pm.name) AS payment_method,
                    GROUP_CONCAT(DISTINCT p.id) AS payment_method_ids
             FROM orders o
+            LEFT JOIN users u ON u.id = o.user_id
             LEFT JOIN tickets t ON t.order_id = o.id
             LEFT JOIN ticket_payments tp ON tp.ticket_id = t.id
             LEFT JOIN payments p ON p.id = tp.payment_id
@@ -85,7 +115,6 @@ class OrderService(
             id = resultSet.getString("id"),
             user_id = resultSet.getString("user_id"),
             table_id = resultSet.getString("table_id"),
-            waiter = resultSet.getString("waiter"),
             status = resultSet.getString("status"),
             total = resultSet.getDouble("total"),
             created_at = resultSet.getString("created_at"),
@@ -100,8 +129,8 @@ class OrderService(
         return OrderWithPayment(
             id = resultSet.getString("id"),
             user_id = resultSet.getString("user_id"),
+            user_name = resultSet.getString("user_name"),
             table_id = resultSet.getString("table_id"),
-            waiter = resultSet.getString("waiter"),
             status = resultSet.getString("status"),
             total = resultSet.getDouble("total"),
             created_at = resultSet.getString("created_at"),
@@ -170,16 +199,15 @@ class OrderService(
         statement.setString(1, generatedId)
         statement.setString(2, order.user_id)
         statement.setString(3, order.table_id)
-        statement.setString(4, order.waiter)
-        statement.setString(5, orderStatus)
-        statement.setDouble(6, order.total)
+        statement.setString(4, orderStatus)
+        statement.setDouble(5, order.total)
         val createdAt =
             order.created_at.ifEmpty {
                 java.time.LocalDateTime
                     .now()
                     .toString()
             }
-        statement.setString(7, createdAt)
+        statement.setString(6, createdAt)
 
         val rowsAffected = statement.executeUpdate()
 
@@ -383,10 +411,9 @@ class OrderService(
         val statement = connection.prepareStatement(UPDATE_ORDER)
         statement.setString(1, order.user_id)
         statement.setString(2, order.table_id)
-        statement.setString(3, order.waiter)
-        statement.setString(4, orderStatus)
-        statement.setDouble(5, order.total)
-        statement.setString(6, order.id)
+        statement.setString(3, orderStatus)
+        statement.setDouble(4, order.total)
+        statement.setString(5, order.id)
 
         val rowsUpdated = statement.executeUpdate()
         if (rowsUpdated > 0) {
@@ -459,5 +486,129 @@ class OrderService(
         val order = getOrderById(orderId) ?: return false
         val updatedOrder = order.copy(total = newTotal)
         return updateOrder(updatedOrder)
+    }
+
+    // ── Store orders ─────────────────────────────────────────────────────────
+
+    private fun mapStoreItems(orderId: String): List<StoreOrderItem> {
+        val st = connection.prepareStatement(STORE_GET_ITEMS)
+        st.setString(1, orderId)
+        val rs = st.executeQuery()
+        val items = mutableListOf<StoreOrderItem>()
+        while (rs.next()) {
+            items.add(
+                StoreOrderItem(
+                    product_id = rs.getString("product_id"),
+                    quantity = rs.getInt("quantity"),
+                    price_at_order = rs.getInt("price_at_order"),
+                ),
+            )
+        }
+        return items
+    }
+
+    private fun mapStoreOrder(rs: java.sql.ResultSet): StoreOrder {
+        val id = rs.getString("id")
+        return StoreOrder(
+            id = id,
+            user_id = rs.getString("user_id"),
+            user_name = rs.getString("user_name"),
+            status = rs.getString("status"),
+            total = rs.getInt("total"),
+            created_at = rs.getString("created_at"),
+            items = mapStoreItems(id),
+        )
+    }
+
+    suspend fun getStoreOrders(status: String? = null): List<StoreOrder> {
+        val sql = if (status != null) STORE_GET_ORDERS_BY_STATUS else STORE_GET_ORDERS
+        val st = connection.prepareStatement(sql)
+        if (status != null) st.setString(1, status)
+        val rs = st.executeQuery()
+        val orders = mutableListOf<StoreOrder>()
+        while (rs.next()) orders.add(mapStoreOrder(rs))
+        return orders
+    }
+
+    suspend fun getStoreOrderById(id: String): StoreOrder? {
+        val st = connection.prepareStatement(STORE_GET_ORDER)
+        st.setString(1, id)
+        val rs = st.executeQuery()
+        return if (rs.next()) mapStoreOrder(rs) else null
+    }
+
+    suspend fun cancelStoreOrder(id: String): Boolean {
+        val st = connection.prepareStatement(STORE_CANCEL_ORDER)
+        st.setString(1, id)
+        val rows = st.executeUpdate()
+        if (rows > 0) logger.info("Store order cancelled: $id")
+        return rows > 0
+    }
+
+    suspend fun checkout(request: StoreCheckoutRequest): StoreCheckoutResponse? {
+        if (request.items.isEmpty()) return null
+        if (request.items.any { it.quantity <= 0 }) return null
+
+        val prev = connection.autoCommit
+        connection.autoCommit = false
+        try {
+            val orderId = UUID.randomUUID().toString()
+            val orderSt = connection.prepareStatement(STORE_INSERT_CHECKOUT_ORDER)
+            orderSt.setString(1, orderId)
+            orderSt.setString(2, request.user_id)
+            orderSt.setDouble(3, request.amount)
+            orderSt.executeUpdate()
+
+            val itemSt = connection.prepareStatement(STORE_INSERT_ORDER_ITEM)
+            val stockSt = connection.prepareStatement(STORE_DECREMENT_STOCK)
+            for (item in request.items) {
+                itemSt.setString(1, orderId)
+                itemSt.setString(2, item.product_id)
+                itemSt.setInt(3, item.quantity)
+                itemSt.setInt(4, item.price_at_order)
+                itemSt.executeUpdate()
+
+                stockSt.setInt(1, item.quantity)
+                stockSt.setString(2, item.product_id)
+                stockSt.setInt(3, item.quantity)
+                val rows = stockSt.executeUpdate()
+                if (rows == 0) {
+                    connection.rollback()
+                    return null
+                }
+            }
+
+            val ticketId = UUID.randomUUID().toString()
+            val ticketSt = connection.prepareStatement(STORE_INSERT_TICKET)
+            ticketSt.setString(1, ticketId)
+            ticketSt.setString(2, orderId)
+            ticketSt.setString(3, request.user_id)
+            ticketSt.setDouble(4, request.amount)
+            ticketSt.setString(5, request.ticket_notes)
+            ticketSt.executeUpdate()
+
+            val paymentId = UUID.randomUUID().toString()
+            val paymentSt = connection.prepareStatement(STORE_INSERT_PAYMENT)
+            paymentSt.setString(1, paymentId)
+            paymentSt.setString(2, request.payment_method_id)
+            paymentSt.setString(3, request.currency_id)
+            paymentSt.setString(4, request.transaction_id ?: "")
+            paymentSt.setDouble(5, request.amount)
+            paymentSt.executeUpdate()
+
+            val tpSt = connection.prepareStatement(STORE_INSERT_TICKET_PAYMENT)
+            tpSt.setString(1, paymentId)
+            tpSt.setString(2, ticketId)
+            tpSt.executeUpdate()
+
+            connection.commit()
+            logger.info("Store checkout: order=$orderId ticket=$ticketId payment=$paymentId")
+            return StoreCheckoutResponse(orderId, ticketId, paymentId)
+        } catch (e: Exception) {
+            connection.rollback()
+            throw e
+        } finally {
+            connection.autoCommit = prev
+        }
     }
 }
