+++ server/app/src/main/resources/db/migration/V24__drop_waiter_from_orders.sql
@@ -1,0 +1,1 @@
+ALTER TABLE orders DROP COLUMN waiter;
